Cleanups, minor fixes, efficiency nits
- (Android) Remove retroarch-jni library - only used for extractArchiveTo
- (Android) Add StorageInterface class with extractArchive() to replace retroarch-jni
- (Android) Install Assets: Show failure toast if no assets found to install
- (Android) Show full paths of storage volumes in launcher
- (Android) Use StringBuilder in getVolumePaths()
- Avoid division in eightway state functions (mainline RetroArch backport)
- nits

+++ audio/audio_driver.c
@@ -837,7 +837,7 @@ void audio_monitor_adjust_system_rates(void)
       return;
 
    timing_ratio       = video_config_get_target_fps() / info->fps;
-   timing_skew        = fabs(1.0f - timing_ratio);
+   timing_skew        = fabsf(1.0f - timing_ratio);
    audio_data.in_rate = info->sample_rate;
 
    if (timing_skew <= settings->audio.max_timing_skew)
